ported version 1.0.1 to singlePrecision
this includes the commits:
----------------------------------------------------------------------
From 7dab661bfc959e4258891591033b3ec536fe68d9 Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Tue, 28 Feb 2012 20:31:33 +0000
Subject: [PATCH 1/5] change: solveDynamicEq should now have better
 performance.

This function contained a loop in which a "current" and "next" set of
coordinates was calculated. This was redundant, as the "next" set of data
at iteration i is the "current" set at i+1.

----------------------------------------------------------------------
From 573aedb996bce6cf581a3ec8565af774e1340baa Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Thu, 1 Mar 2012 15:23:47 +0000
Subject: [PATCH 2/5] change: converted all functions from double precision to
 single precision

With the exception of those that are matlab-related as matlab formats use
double precision internally.

----------------------------------------------------------------------
From 7d7dadb117bbef031764a5f9bf16c24f7205a7f8 Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Fri, 9 Mar 2012 15:57:31 +0000
Subject: [PATCH 3/5] change: created a test case to compare Single Precision
 to DP version.

Also added a tag to the version identifier to make it easy to recognize
distinguish between SP and DP.

----------------------------------------------------------------------
From d6db203f768eb237d17a18bd97b0b754743ae371 Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Tue, 13 Mar 2012 15:08:16 +0000
Subject: [PATCH 4/5] change: when trying to use parabolic interpolation a
 warning is shown.

Since parabolic interpolation is no longer supported (because it can generate
incorrect results), a warning is now shown when reading the input file.
Instead of 3P interpolation, 4P will be used and the warning will be shown.

----------------------------------------------------------------------
From a241064073cbfd52511b26dc7a2e0ace9202d7cd Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Tue, 13 Mar 2012 15:31:53 +0000
Subject: [PATCH 5/5] change: added warnings about precision when using
 linear/2P interpolation

+++ source/boundaryInterpolation.c
@@ -50,14 +50,14 @@
 
 
 
-void boundaryInterpolationExplicit(uint32_t*, double*, double*, uint32_t*, double, double*, vector_t*, vector_t*);
-void boundaryInterpolation(interface_t*, double, double*, vector_t*, vector_t*);
+void boundaryInterpolationExplicit(uint32_t*, float*, float*, uint32_t*, float, float*, vector_t*, vector_t*);
+void boundaryInterpolation(interface_t*, float, float*, vector_t*, vector_t*);
 
-void boundaryInterpolationExplicit(uint32_t* numSurfaceCoords, double* r, double* z, uint32_t* surfaceInterpolation, double ri, double* zi, vector_t* taub, vector_t* normal){
+void boundaryInterpolationExplicit(uint32_t* numSurfaceCoords, float* r, float* z, uint32_t* surfaceInterpolation, float ri, float* zi, vector_t* taub, vector_t* normal){
     DEBUG(5,"in\n");
         
-    double      zri = 0;    //1st derivative of z at ri
-    double      zrri;   //2nd derivative of z at ri
+    float      zri = 0;    //1st derivative of z at ri
+    float      zrri;   //2nd derivative of z at ri
     uintptr_t   i = 0;
     
     switch(*surfaceInterpolation){
@@ -119,7 +119,7 @@ void boundaryInterpolationExplicit(uint32_t* numSurfaceCoords, double* r, double
     DEBUG(5,"out\n");
 }
 
-void boundaryInterpolation(interface_t* interface, double ri, double* zi, vector_t* taub, vector_t* normal){
+void boundaryInterpolation(interface_t* interface, float ri, float* zi, vector_t* taub, vector_t* normal){
     boundaryInterpolationExplicit(  &(interface->numSurfaceCoords),
                                     interface->r,
                                     interface->z,
