+++ source/dotProduct.c
@@ -42,8 +42,9 @@
 
 #pragma once
 
-double  dotProduct(vector_t*, vector_t*);
+float  dotProduct(vector_t*, vector_t*);
 
-double  dotProduct(vector_t* u, vector_t* v){
+float  dotProduct(vector_t* u, vector_t* v){
+    //TODO: replace this with a call to openCL's dotProduct
     return( u->r * v->r + u->z * v->z);
 }
