+++ source/intComplexLinear1D.c
@@ -44,9 +44,9 @@
 
 #pragma once
 #include <complex.h>
-void intComplexLinear1D(double*, complex double*, complex double, complex double*, complex double*);
+void intComplexLinear1D(float*, complex float*, complex float, complex float*, complex float*);
 
-void intComplexLinear1D(double* x, complex double* f, complex double xi, complex double* fi, complex double* fxi){
+void intComplexLinear1D(float* x, complex float* f, complex float xi, complex float* fi, complex float* fxi){
     DEBUG(8,"in\n");
     *fxi    =   ( f[1] -f[0]) / ( x[1] -x[0]);
     *fi     =   f[0] +(xi -x[0]) *(*fxi);
