+++ source/pressureMStar.c
@@ -60,17 +60,17 @@
 #include <complex.h>
 #include "eBracket.c"
 
-uintptr_t   pressureMStar(settings_t*, ray_t*, double, double, double, complex double*, complex double[]);
+uintptr_t   pressureMStar(settings_t*, ray_t*, float, float, float, complex float*, complex float[]);
 
-uintptr_t   pressureMStar( settings_t* settings, ray_t* ray, double rHyd, double zHyd, double q0, complex double* pressure_H, complex double* pressure_V){
+uintptr_t   pressureMStar( settings_t* settings, ray_t* ray, float rHyd, float zHyd, float q0, complex float* pressure_H, complex float* pressure_V){
 
-    double              rLeft, rRight, zTop, zBottom;
+    float              rLeft, rRight, zTop, zBottom;
     uintptr_t           i, jj;
-    double              dzdr, tauRay, zRay, qRay, width;
-    complex double      ampRay;
+    float              dzdr, tauRay, zRay, qRay, width;
+    complex float      ampRay;
     uintptr_t           nRet;
     uintptr_t           iRet[51];
-    complex double      tempPressure[3];
+    complex float      tempPressure[3];
     
     /* start with determining the coordinates for which we will need to calculate
      * acoustic pressure.
