+++ source/rayBoundaryIntersection.c
@@ -85,9 +85,9 @@ void    rayObjectIntersection(objects_t* objects, uint32_t* j, int32_t boundary,
 void    rayBoundaryIntersection(interface_t* interface, point_t* a, point_t* b, point_t* isect){
     DEBUG(4,"in\n");
     uint32_t    i,n;
-    double      rl[101];
-    double      zl[101];
-    double      dz[101];
+    float      rl[101];
+    float      zl[101];
+    float      dz[101];
     point_t     q1;
     point_t     q2;
     vector_t    taub;
@@ -138,7 +138,7 @@ void    rayBoundaryIntersection(interface_t* interface, point_t* a, point_t* b, 
                 isect->z = 0.0;
             }
             break;
-//TODO double check the last 20~ lines
+//TODO float check the last 20~ lines
     }
     DEBUG(4,"out\n");
 }
